Доработан функционал Comments

+++ main-service/src/main/java/ru/practicum/comment/service/CommentServiceImpl.java
@@ -6,19 +6,23 @@ import org.springframework.transaction.annotation.Transactional;
 import ru.practicum.comment.dto.AdminUpdateCommentStatusDto;
 import ru.practicum.comment.dto.CommentDto;
 import ru.practicum.comment.dto.NewCommentDto;
-import ru.practicum.comment.dto.UpdateCommentDto;
 import ru.practicum.comment.enums.AdminUpdateCommentStatusAction;
 import ru.practicum.comment.enums.CommentStatus;
-import ru.practicum.comment.enums.UpdateCommentAction;
 import ru.practicum.comment.mapper.CommentMapper;
 import ru.practicum.comment.model.Comment;
 import ru.practicum.comment.repository.CommentRepository;
 import ru.practicum.common.exception.NotFoundException;
 import ru.practicum.common.exception.OperationForbiddenException;
 import ru.practicum.events.model.Event;
+import ru.practicum.events.model.EventState;
 import ru.practicum.events.repository.EventRepository;
+import ru.practicum.request.model.RequestStatus;
+import ru.practicum.request.repository.RequestRepository;
 import ru.practicum.user.model.User;
 import ru.practicum.user.repository.UserRepository;
+
+import java.time.LocalDateTime;
+import java.util.List;
 
 @Service
 @RequiredArgsConstructor
@@ -28,14 +32,24 @@ public class CommentServiceImpl implements CommentService {
     private final UserRepository userRepository;
     private final EventRepository eventRepository;
     private final CommentMapper commentMapper;
+    private final RequestRepository requestRepository;
 
     @Transactional
     @Override
     public CommentDto createComment(long authorId, long eventId, NewCommentDto newCommentDto) {
         User author = userRepository.findById(authorId)
                 .orElseThrow(() -> new NotFoundException(String.format("User with ID %s not found", authorId)));
-        Event event = eventRepository.findByIdAndInitiator_Id(eventId, authorId)
+        Event event = eventRepository.findById(eventId)
                 .orElseThrow(() -> new NotFoundException(String.format("Event with ID %s not found", eventId)));
+        if (authorId == event.getInitiator().getId()) {
+            throw new OperationForbiddenException("Инициатор мероприятия не может оставлять комментарии к нему");
+        }
+        if (!event.getState().equals(EventState.PUBLISHED)||!event.getEventDate().isBefore(LocalDateTime.now())) {
+            throw new OperationForbiddenException("Мероприятие должно быть опубликовано, а дата его проведения в прошлом");
+        }
+        if (requestRepository.findByRequesterIdAndEventIdAndStatus(authorId, eventId, RequestStatus.CONFIRMED).isEmpty()) {
+            throw new OperationForbiddenException("Комментарии может оставлять только подтвержденный участник мероприятия");
+        }
         Comment comment = commentMapper.toComment(newCommentDto, author, event);
         commentRepository.save(comment);
         return commentMapper.toDto(comment);
@@ -43,40 +57,27 @@ public class CommentServiceImpl implements CommentService {
 
     @Transactional
     @Override
-    public CommentDto updateComment(long authorId, long eventId, long commentId, UpdateCommentDto updateCommentDto) {
-        User author = userRepository.findById(authorId)
-                .orElseThrow(() -> new NotFoundException(String.format("User with ID %s not found", authorId)));
-        Event event = eventRepository.findByIdAndInitiator_Id(eventId, authorId)
-                .orElseThrow(() -> new NotFoundException(String.format("Event with ID %s not found", eventId)));
+    public CommentDto updateComment(long authorId, long commentId, NewCommentDto updateCommentDto) {
         Comment commentToUpdate = commentRepository.findById(commentId)
                 .orElseThrow(() -> new NotFoundException(String.format("Comment with ID %s not found", commentId)));
-        if (updateCommentDto.getText() != null) {
-            commentToUpdate.setText(updateCommentDto.getText());
+        if (authorId != commentToUpdate.getAuthor().getId()) {
+            throw new OperationForbiddenException("Изменить комментарий может только его автор");
         }
-        if (updateCommentDto.getAction() != null) {
-            if (updateCommentDto.getAction().equals(UpdateCommentAction.SEND_TO_REVIEW)) {
-                commentToUpdate.setStatus(CommentStatus.PENDING);
-            }
-            if (updateCommentDto.getAction().equals(UpdateCommentAction.CANCEL_REVIEW)) {
-                commentToUpdate.setStatus(CommentStatus.CANCELED);
-        } else {
-                commentToUpdate.setStatus(CommentStatus.PENDING);
-            }
+        commentToUpdate.setText(updateCommentDto.getText());
+        commentToUpdate.setStatus(CommentStatus.PENDING);
 
-        }
         commentRepository.save(commentToUpdate);
         return commentMapper.toDto(commentToUpdate);
     }
 
     @Transactional
     @Override
-    public void deleteComment(long authorId, long eventId, long commentId) {
-        User author = userRepository.findById(authorId)
-                .orElseThrow(() -> new NotFoundException(String.format("User with ID %s not found", authorId)));
-        Event event = eventRepository.findByIdAndInitiator_Id(eventId, authorId)
-                .orElseThrow(() -> new NotFoundException(String.format("Event with ID %s not found", eventId)));
+    public void deleteComment(long authorId, long commentId) {
         Comment commentToDelete = commentRepository.findById(commentId)
                 .orElseThrow(() -> new NotFoundException(String.format("Comment with ID %s not found", commentId)));
+        if (authorId != commentToDelete.getAuthor().getId()) {
+            throw new OperationForbiddenException("Удалить комментарий может только его автор");
+        }
         commentRepository.delete(commentToDelete);
     }
 
@@ -96,5 +97,13 @@ public class CommentServiceImpl implements CommentService {
         }
         commentRepository.save(commentToUpdateStatus);
         return commentMapper.toDto(commentToUpdateStatus);
+    }
+
+    @Override
+    public List<CommentDto> adminPendigCommentList() {
+        return commentRepository.findAllByStatus(CommentStatus.PENDING)
+                .stream()
+                .map(commentMapper::toDto)
+                .toList();
     }
 }
